backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
chromadb
sqlalchemy
sentence-transformers
pandas
numpy
scikit-learn
xgboost
matplotlib
plotly
PyMuPDF
httpx
python-dotenv
aiofiles
python-multipart
openpyxl
sse-starlette
